# tests/yoke/http/golden/openapi.json
{
  "components": {
    "schemas": {
...
      "CursorInfo": {
        "additionalProperties": false,
        "description": "Opaque list pagination cursors.",
        "properties": {
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous"
          }
        },
        "title": "CursorInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ToolCallInfo": {
        "additionalProperties": false,
        "properties": {
          "afterContext": {
            "items": {
              "$ref": "#/components/schemas/ToolTraceContextInfo"
            },
            "title": "Aftercontext",
            "type": "array"
          },
          "arguments": {
            "$ref": "#/components/schemas/ToolTraceArguments"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/ToolTraceContextInfo"
            },
            "title": "Context",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "iteration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iteration"
          },
          "output": {
            "$ref": "#/components/schemas/ToolTraceOutputInfo"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "retention": {
            "enum": [
              "runtime",
              "session"
            ],
            "title": "Retention",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "ok",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "time": {
            "$ref": "#/components/schemas/ToolTraceTime"
          },
          "toolName": {
            "title": "Toolname",
            "type": "string"
          },
          "turnID": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnid"
          }
        },
        "required": [
          "id",
          "toolName",
          "status",
          "arguments",
          "time",
          "output",
          "context",
          "afterContext",
          "retention"
        ],
        "title": "ToolCallInfo",
        "type": "object"
      },
      "ToolCallListResponse": {
        "additionalProperties": false,
        "properties": {
          "cursor": {
            "$ref": "#/components/schemas/CursorInfo"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ToolCallInfo"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data",
          "cursor"
        ],
        "title": "ToolCallListResponse",
        "type": "object"
      },
      "ToolCallResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ToolCallInfo"
          }
        },
        "required": [
          "data"
        ],
        "title": "ToolCallResponse",
        "type": "object"
      },
...
      "ToolOutputChunk": {
        "additionalProperties": false,
        "properties": {
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "stream": {
            "title": "Stream",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "stream",
          "text"
        ],
        "title": "ToolOutputChunk",
        "type": "object"
      },
      "ToolOutputCursor": {
        "additionalProperties": false,
        "properties": {
          "next": {
            "title": "Next",
            "type": "integer"
          },
          "truncatedBefore": {
            "title": "Truncatedbefore",
            "type": "integer"
          }
        },
        "required": [
          "next",
          "truncatedBefore"
        ],
        "title": "ToolOutputCursor",
        "type": "object"
      },
      "ToolOutputResponse": {
        "additionalProperties": false,
        "properties": {
          "cursor": {
            "$ref": "#/components/schemas/ToolOutputCursor"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ToolOutputChunk"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data",
          "cursor"
        ],
        "title": "ToolOutputResponse",
        "type": "object"
      },
...
      "ToolTraceArguments": {
        "additionalProperties": false,
        "properties": {
          "executed": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed"
          },
          "raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          }
        },
        "title": "ToolTraceArguments",
        "type": "object"
      },
      "ToolTraceContextInfo": {
        "additionalProperties": false,
        "properties": {
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "role",
          "text"
        ],
        "title": "ToolTraceContextInfo",
        "type": "object"
      },
      "ToolTraceOutputInfo": {
        "additionalProperties": false,
        "properties": {
          "latestSeq": {
            "default": 0,
            "title": "Latestseq",
            "type": "integer"
          },
          "retainedChars": {
            "default": 0,
            "title": "Retainedchars",
            "type": "integer"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "title": "ToolTraceOutputInfo",
        "type": "object"
      },
      "ToolTraceTime": {
        "additionalProperties": false,
        "properties": {
          "durationMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationms"
          },
          "ended": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended"
          },
          "started": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started"
          }
        },
        "title": "ToolTraceTime",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Yoke HTTP API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/session/{session_id}/tool-call": {
      "get": {
        "operationId": "listSessionToolCalls",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "running",
                    "ok",
                    "failed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "turnID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Turnid"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCallListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tool Calls"
      }
    },
    "/api/v1/session/{session_id}/tool-call/{call_id}": {
      "get": {
        "operationId": "getSessionToolCall",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "title": "Call Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tool Call"
      }
    },
    "/api/v1/session/{session_id}/tool-call/{call_id}/output": {
      "get": {
        "operationId": "getSessionToolCallOutput",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "title": "Call Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "afterSeq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Afterseq",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolOutputResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tool Output"
      }
    },
...
  }
}